tweak SysProps
